revert scoreDoc, move html generation to core
